[test]: CommandTest changed due to moving some commands to waiting to edit file.
