+++ src/test/java/systemtests/ClearCommandSystemTest.java
@@ -6,10 +6,9 @@ import static seedu.address.testutil.TypicalPersons.KEYWORD_MATCHING_MEIER;
 import org.junit.Test;
 
 import seedu.address.commons.core.index.Index;
-import seedu.address.logic.commands.ClearCommand;
-import seedu.address.logic.commands.RedoCommand;
-import seedu.address.logic.commands.UndoCommand;
-import seedu.address.model.Model;
+import seedu.address.logic.command.editing.ClearCommand;
+import seedu.address.logic.command.editing.RedoCommand;
+import seedu.address.logic.command.editing.UndoCommand;
 import seedu.address.model.ModelManager;
 
 public class ClearCommandSystemTest extends AddressBookSystemTest {
