+++ src/test/java/systemtests/EditCommandSystemTest.java
@@ -35,10 +35,9 @@ import org.junit.Test;
 
 import seedu.address.commons.core.Messages;
 import seedu.address.commons.core.index.Index;
-import seedu.address.logic.commands.EditCommand;
-import seedu.address.logic.commands.RedoCommand;
-import seedu.address.logic.commands.UndoCommand;
-import seedu.address.model.Model;
+import seedu.address.logic.command.editing.EditCommand;
+import seedu.address.logic.command.editing.RedoCommand;
+import seedu.address.logic.command.editing.UndoCommand;
 import seedu.address.model.person.Address;
 import seedu.address.model.person.Email;
 import seedu.address.model.person.Name;
