+++ src/test/java/systemtests/FindCommandSystemTest.java
@@ -14,11 +14,10 @@ import java.util.List;
 import org.junit.Test;
 
 import seedu.address.commons.core.index.Index;
-import seedu.address.logic.commands.DeleteCommand;
-import seedu.address.logic.commands.FindCommand;
-import seedu.address.logic.commands.RedoCommand;
-import seedu.address.logic.commands.UndoCommand;
-import seedu.address.model.Model;
+import seedu.address.logic.command.editing.DeleteCommand;
+import seedu.address.logic.command.editing.FindCommand;
+import seedu.address.logic.command.editing.RedoCommand;
+import seedu.address.logic.command.editing.UndoCommand;
 import seedu.address.model.tag.Tag;
 
 public class FindCommandSystemTest extends AddressBookSystemTest {
