+++ src/test/java/systemtests/HelpCommandSystemTest.java
@@ -11,9 +11,9 @@ import org.junit.Test;
 
 import guitests.GuiRobot;
 import guitests.guihandles.HelpWindowHandle;
-import seedu.address.logic.commands.DeleteCommand;
-import seedu.address.logic.commands.HelpCommand;
-import seedu.address.logic.commands.SelectCommand;
+import seedu.address.logic.command.editing.DeleteCommand;
+import seedu.address.logic.command.editing.HelpCommand;
+import seedu.address.logic.command.editing.SelectCommand;
 import seedu.address.ui.BrowserPanel;
 import seedu.address.ui.StatusBarFooter;
